cleanup index.ts imports / exports
